fix: revert enable/disable via SIDEKIQ_CRON_ENABLE flag
We discovered from last Friday's incident that the SIDEKIQ_CRON_
ENABLE flag is in practical usage overshadowed by SIDEKIQ_STARTUP_
CHECKS being set to 0, because disabling Sidekiq's scheduler also
disables sidekiq-cron from launching (thanks to OOP inheritance).
This commit is to undo that addition from PR #1 as we're not using
the flag in any meaningful way.